+++ src/lib/pcm_write.c
@@ -109,7 +109,7 @@ pcm_write_i16le(
 /*@modifies	*dest@*/
 {
 	register size_t i, j;
-	for ( i = 0, j = 0; i < nsamples; ++i, j += 2u ){
+	for ( i = 0, j = 0; i < nsamples; ++i, j += (size_t) 2u ){
 		write_i32h_to_i16le(&dest[j], src[i]);
 	}
 	return i;
@@ -124,7 +124,7 @@ pcm_write_i24le(
 /*@modifies	*dest@*/
 {
 	register size_t i, j;
-	for ( i = 0, j = 0; i < nsamples; ++i, j += 3u ){
+	for ( i = 0, j = 0; i < nsamples; ++i, j += (size_t) 3u ){
 		write_i32h_to_i24le(&dest[j], src[i]);
 	}
 	return i;
@@ -144,8 +144,8 @@ write_i32h_to_i16le(register u8 *const dest, register i32 x)
 /*@modifies	*dest@*/
 {
 	register const u32 y = (u32) x;
-	dest[0] = (u8)  y;
-	dest[1] = (u8) (y >> 8u);
+	dest[0u] = (u8)  y;
+	dest[1u] = (u8) (y >> 8u);
 	return;
 }
 
@@ -154,9 +154,9 @@ write_i32h_to_i24le(register u8 *const dest, register i32 x)
 /*@modifies	*dest@*/
 {
 	register const u32 y = (u32) x;
-	dest[0] = (u8)  y;
-	dest[1] = (u8) (y >>  8u);
-	dest[2] = (u8) (y >> 16u);
+	dest[0u] = (u8)  y;
+	dest[1u] = (u8) (y >>  8u);
+	dest[2u] = (u8) (y >> 16u);
 	return;
 }
